backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.8.2
python-multipart==0.0.9
